rename some structlog files and classes for clarity
