+++ prefab_cloud_python/structlog_multi_processor.py
@@ -1,4 +1,4 @@
-class MultiProcessor:
+class StructlogMultiProcessor:
     def __init__(self, processors):
         self.processors = processors
